frame: moved fuzzing tests to distinct files
