+++ frame/response/authsuccess_test.go
@@ -34,16 +34,3 @@ func TestAuthSuccessEncodeDecode(t *testing.T) {
 		})
 	}
 }
-
-var dummyAS *AuthSuccess
-
-// We want to make sure that parsing does not crush driver even for random data.
-// We assign result to global variable to avoid compiler optimization.
-func FuzzAuthSuccess(f *testing.F) {
-	f.Fuzz(func(t *testing.T, data []byte) { // nolint:thelper // This is not a helper function.
-		var buf frame.Buffer
-		buf.Write(data)
-		out := ParseAuthSuccess(&buf)
-		dummyAS = out
-	})
-}
