+++ frame/response/supported_test.go
@@ -95,16 +95,3 @@ func TestParseScyllaSupported(t *testing.T) {
 		})
 	}
 }
-
-var dummyS *Supported
-
-// We want to make sure that parsing does not crush driver even for random data.
-// We assign result to global variable to avoid compiler optimization.
-func FuzzSupported(f *testing.F) {
-	f.Fuzz(func(t *testing.T, data []byte) { // nolint:thelper // This is not a helper function.
-		var buf frame.Buffer
-		buf.Write(data)
-		out := ParseSupported(&buf)
-		dummyS = out
-	})
-}
